The MicrobiomeR Formats

The formats are listed in order from lowest to highest level. To get a taxmap object into the higher formats they must be in one of the lower formats. It is also inherently important to understand the MicrobiomeR_Workflow, which describes some of the important dependencies used to create the MicrobiomeR formats listed below:

phyloseq_format

Level 0. The "phyloseq_format" is the format given to a taxmap object that has just been converted from a phyloseq object using metacoder via create_phyloseq. A taxmap object identified by MicrobiomeR to be in this format will contain "otu_table", "tax_data", "sample_data" and "phy_tree" data in the observation tables (metacoder_obj$data), which represents the matching phyloseq::otu_table, phyloseq::tax_table, phyloseq::sample_data, and phyloseq::phy_tree data in the original phyloseq object.

raw_format

Level 1. The "raw_format" is the format given to a taxmap object that has been processed with as_raw_format. It contains the "otu_abundance" and "otu_annotations" data in the observation tables, which are just name conversions of the otu_table and tax_data table from the "phyloseq_format".

basic_format

Level 2. The "basic_format" is the format given to a taxmap object that has been processed with as_basic_format. It contains the tables available in the "raw_format" along with "taxa_abundance", "otu_proportions", and "taxa_proportions" tables.

analyzed_format

Level 3. The "analyzed_format" is the format given to a taxmap object that has been processed with as_analyzed_format. It contains the tables available in the "basic_format" along with "statistical_data" and "stats_tax_data" tables.

See also