Get Phyloseq Object

Create a phyloseq object using a biom file, phylogenetic tree file, and a metadata file. Alternatively, an .Rdata file can be used. For now this funciton takes output from Qiime. The data provided by this package was processed by the NIH's Nephele pipeline.

get_phyloseq_obj(biom_file = NULL, tree_file = NULL,
  metadata_file = NULL, parse_func = NULL, rdata_file = NULL,
  path = NULL)

Arguments

biom_file
A file in the BIOM format. Default: NULL

tree_file
A Phylogenetic tree file. Default: NULL

metadata_file
Sample metadata in tab delimited format. Default: NULL

parse_func
The parse function used to parse taxonomy strings from Greengenes or SILVA database. Default: NULL

rdata_file
A .Rdata file. Default: NULL

path
A path for the output tree file. Default: NULL

Value

A phyloseq object, and a phylogenetic tree file if one does not already exist.

Details

This function heavily relys on the phyloseq package to import data into R. It also requires you to use an absolute path to your data files or for your data files to be in the working directory.

See also

Examples

# NOT RUN {
> biom_file <- "input_data/silva_OTU.biom"
> md_file <- "input_data/nephele_metadata.txt"
> phy_obj <- get_phyloseq_obj(biom_file=biom_file, metadata_file=md_file)
# }