MicrobiomeR

A microbiome workflow package using phyloseq for data import, metacoder for data analysis, and ggplot2 for data visualization.

Details

The goal of this package is to enhance the use of the metacoder package by providing tools that mimic the phyloseq package. The only difference being that the main data object is a taxmap object instead of a phyloseq object.

Other Documentation

Below are some links to helpful R package used, while developing this package.

metacoder

https://grunwaldlab.github.io/metacoder_documentation/

taxa

https://github.com/ropensci/taxa

phyloseq

http://joey711.github.io/phyloseq/

microbiome

http://microbiome.github.io/microbiome/

MicrobiomeR

https://github.com/vallenderlab/MicrobiomeR

The MicrobiomeR_Workflow and MicrobiomeR_Formats help sections are also useful for understanding the package.