The MicrobiomeR Workflow

In order to understand the MicrobiomeR packages workflow it is necessary to understand a few other packages. Primarily phyloseq, taxa, and metacoder.

Data Import

Data is imported with the create_phyloseq function or something proprietary that creates a phyloseq object with an phyloseq::otu_table, a phyloseq::tax_table, a phyloseq::sample_data, and a phyloseq::phy_tree. The data is preferrably not filtered with phyloseq at this stage.

Data Manipulation

Data is manipulated into a taxa::taxmap object using the metacoder::parse_phyloseq function. Custom MicrobiomeR_Formats are used to validate the data for consistency. MicrobiomeR and metacoder filtering functions can be used to preprocess the data before visualization.

Visualizations

MicrobiomeR contains useful functions for creating visualizations such as heat_trees, correlations plots, stacked bar charts, violin plots, and PCA plots. These functions are also coupled with save functions that organize saved output by experiment.