Alpha Diversity Plot

Plot the alpha diversity using a violin plot. alpha_diversity_plots generates plots for all alpha diversity measures.

alpha_diversity_plot(
  obj,
  measure = "Shannon",
  group = "TreatmentGroup",
  select_otu_table = NULL,
  title = NULL
)

alpha_diversity_plots(
  obj,
  measures = c("Shannon", "GiniSimpson", "InverseSimpson"),
  group = "TreatmentGroup"
)

Arguments

obj

An object to be converted to a Taxmap object with create_taxmap.

measure

Select an alpha diversity measure such as Shannon, Fisher, Coverage, GiniSimpson, and InverseSimpson, Default: 'Shannon'

group

The "TreatmentGroup" or similar grouping or column from your metadata to denote sample groups, Default: 'TreatmentGroup'

select_otu_table

DEPRECATED. Choose an otu table to analyze, Default: 'otu_proportions'

title

The title of the plot, Default: NULL

measures

A list of alpha diversity measures such as Shannon, Fisher, Coverage, GiniSimpson, and InverseSimpson, Default: 'c("Shannon", "GiniSimpson", "InverseSimpson")'

Value

Returns an alpha diversity plot.

Returns a melted dataframe.

Details

Alpha diversity helps to determine the species richness (the number of different species in a sample) or evenness (similar abundance level). We prefer to use Shannon as it is better for data generated using the QIIME pipeline.

See also

Examples

if (FALSE) { # \dontrun{
if (interactive()) {
  library(MicrobiomeR)
  data <- analyzed_silva
  plot <- alpha_diversity_plot(obj = data,
                               measure = "Shannon")
  plot
}
} # }