Alpha Diversity Plot

Plot the alpha diversity using a violin plot.

alpha_diversity_plot(obj, measure = "shannon",
  select_otu_table = "otu_proportions")

Arguments

obj
An object to be converted to a Taxmap object with create_taxmap.

measure
Select an alpha diversity measure such as shannon, gini simpson, and inverse simpson, Default: 'shannon'

select_otu_table
Choose an otu table to analyze, Default: 'otu_proportions'

Value

Returns an alpha diversity plot.

Details

Alpha diversity helps to determine the species richness (the number of different species in a sample) or evenness (similar abundance level). We prefer to use shannon as it is better for data generated using the QUIIME pipeline.

See also

Examples

# NOT RUN {
if (interactive()) {
  library(MicrobiomeR)
  data <- analyzed_silva
  plot <- alpha_diversity_plot(obj = data, measure = "shannon", select_otu_table = "otu_proportions", save = FALSE)
  plot
}
# }