Get Correlation Plot Data

Get the correlation plot data comparing all of the treatment groups.

correlation_data(obj, primary_rank, secondary_rank = TRUE,
  wp_value = 0.05)

Arguments

obj
An object to be converted to a Taxmap object with create_taxmap.

primary_rank
A primary rank used to label the points.

secondary_rank
The secondary rank used to color the points, Default: TRUE

wp_value
The wilcoxon p-value cutoff/threshold, Default: 0.05

Value

Data for the correlation plots.

Details

DETAILS

See also

Examples

# NOT RUN {
if(interactive()){
 #EXAMPLE1
 }
# }