Get Multiple Correlation Plots

This function allows the user to create a list of multiple correlation plots.

correlation_plots(obj, primary_ranks, secondary_ranks = TRUE, ...)

Arguments

obj

An object to be converted to a Taxmap object with create_taxmap.

primary_ranks

A vector of primary ranks used to label the points.

secondary_ranks

The secondary rank used to color the points. Can be an integer specifying the number of supertaxon ranks above the primary rank or the name of a supertaxon rank. Default: TRUE

...

An optional list of parameters to use in the correlation_plot function.

Value

A list object containing correlation plots. A pairwise comparison returns a nested list.

Details

This function makes it easier to generate multiple correlation plots at once.

See also

Examples

if (FALSE) { # \dontrun{
if(interactive()){
# This example uses data that are no longer available in the MicrobiomeR package,
# however, they can be easily generated with \code{\link{MicrobiomeR}{as_analyzed_format}}.
library(MicrobiomeR)
analyzed_silva <- as_MicrobiomeR_format(MicrobiomeR::raw_silva_2, "analyzed_format")
corr_plots <- correlation_plots(analyzed_silva, primary_ranks = c("Phylum", "Class", "Order"),
                      secondary_ranks = c("Phylum", "Class", "Order", "Family", "Genus"))
# Show a plot
corr_plots$Class$Phylum
 }
} # }