Get Heat Tree Parameters

This function gets the parameters used for the heat_tree_plots function.

heat_tree_parameters(obj, title, treatment_no, ...)

Arguments

obj
A Taxmap object.

title
The title used in the heat_tree plot.

treatment_no
The number of treatment groups in the data.

...
Any of the heat tree parameters list below can be used to change the way the heat_tree output is displayed. However, this function acts as a default list of parameters. The members of the default list will be overridden by the dot parameters. Any variable in obj$data$stats_tax_data can be used to manipulate the heat tree parameters. Function calls from the taxa package must be done explicitly on the Taxmap object.

Value

A list used with do.call and the metacoder::heat_tree function.

See also