Pick Outgroup for Tree

Pick an outgroup for rooting a phylogenetic tree.

pick_new_outgroup(unrooted_tree)

Arguments

unrooted_tree

An unrooted tree object.

Value

A new tree with the longest branch as the outgroup.

Details

This function preprocess a phylogenetic tree for rooting by the longest edge. Please see this issue for more details https://github.com/joey711/phyloseq/issues/597.

See also