Save Correlation Plots

This function saves correlation plots storred in a listlike object to an output folder.

save_correlation_plots(corr, format = "tiff", start_path = "output",
  ...)

Arguments

corr
A correlation plot list generated by correlation_plot or get_correlation_plots.

format
The format of the output image. Default: 'tiff'

start_path
The starting path of the output directory. Default: 'output'

...
An optional list of parameters to use in the get_output_dir function.

Value

An output directory that contains correlation plots.

Details

This function creates an appropriate output directory, where it saves publication ready plots.

See also

Examples

# NOT RUN {
if(interactive()){
# This example uses data that are no longer available in the MicrobiomeR package,
# however, they can be easily generated with \code{\link{MicrobiomeR}{as_analyzed_silva}}.
library(MicrobiomeR)
analyzed_silva <- as_MicrobiomeR_format(MicrobiomeR::raw_silva_2, "analyzed_format")
corr_plot <- correlation_plot(analyzed_silva, primary_rank = "Class", secondary_rank = "Phylum")
# Save to \emph{./output/corr_plot} folder.
save_correlation_plots(corr_plot)
 }
# }